third_party/WebKit/LayoutTests/external/wpt/html/dom/dynamic-markup-insertion/opening-the-input-stream/009-expected.txt

This is a testharness.js-based test.
FAIL document.open replacing singleton windowassert_equals: expected (undefined) undefined but got (number) 1
FAIL document.open replacing singleton locationassert_equals: expected (undefined) undefined but got (number) 1
FAIL document.open replacing singleton historyassert_equals: expected (undefined) undefined but got (number) 1
FAIL document.open replacing singleton navigatorassert_equals: expected (undefined) undefined but got (number) 1
FAIL document.open replacing singleton applicationCacheassert_equals: expected (undefined) undefined but got (number) 1
FAIL document.open replacing singleton sessionStorageassert_equals: expected (undefined) undefined but got (string) "1"
FAIL document.open replacing singleton localStorageassert_equals: expected (undefined) undefined but got (string) "1"
FAIL document.open replacing singleton locationbarassert_equals: expected (undefined) undefined but got (number) 1
Harness: the test ran to completion.
